requirements.txt
Pillow~=6.2.1
lark-parser~=0.7
# pytest~=5.0
